# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Capacity & Spend Control API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Sign up a new user (first user becomes Admin)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SignupInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MeResponse" }
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MeResponse" }
...
  /allocations:
    get:
      operationId: listAllocations
      tags: [allocations]
      parameters:
        - { name: projectId, in: query, schema: { type: integer } }
        - { name: resourceId, in: query, schema: { type: integer } }
        - { name: year, in: query, schema: { type: integer } }
        - { name: fy, in: query, schema: { type: integer }, description: "Fiscal year (Jul–Jun, named by ending year)" }
        - { name: month, in: query, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Allocation" }
    post:
      operationId: createAllocation
      tags: [allocations]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AllocationInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Allocation" }
  /allocations/{id}:
    patch:
      operationId: updateAllocation
      tags: [allocations]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AllocationInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Allocation" }
    delete:
      operationId: deleteAllocation
      tags: [allocations]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: Deleted }
...
  /actuals:
    get:
      operationId: listActuals
      tags: [actuals]
      parameters:
        - { name: projectId, in: query, schema: { type: integer } }
        - { name: year, in: query, schema: { type: integer } }
        - { name: fy, in: query, schema: { type: integer }, description: "Fiscal year (Jul–Jun, named by ending year)" }
        - { name: source, in: query, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Actual" }
    post:
      operationId: createActual
      tags: [actuals]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ActualInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Actual" }
  /actuals/{id}:
    patch:
      operationId: updateActual
      tags: [actuals]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ActualInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Actual" }
    delete:
      operationId: deleteActual
      tags: [actuals]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: Deleted }
...
components:
  schemas:
...
    Allocation:
      type: object
      required:
        - id
        - projectId
        - projectName
        - resourceId
        - resourceName
        - month
        - year
        - plannedDays
        - allocationPct
        - rate
        - plannedCost
        - opexPct
        - capexPct
        - wbsCode
      properties:
        id: { type: integer }
        projectId: { type: integer }
        projectName: { type: string }
        resourceId: { type: integer }
        resourceName: { type: string }
        month: { type: integer, minimum: 1, maximum: 12 }
        year: { type: integer }
        plannedDays: { type: number }
        allocationPct: { type: number }
        rate: { type: number }
        plannedCost: { type: number }
        opexPct: { type: number }
        capexPct: { type: number }
        wbsCode: { type: string }
    AllocationInput:
      type: object
      required:
        - projectId
        - resourceId
        - month
        - year
        - plannedDays
        - allocationPct
        - rate
        - opexPct
        - capexPct
        - wbsCode
      properties:
        projectId: { type: integer }
        resourceId: { type: integer }
        month: { type: integer }
        year: { type: integer }
        plannedDays: { type: number }
        allocationPct: { type: number }
        rate: { type: number }
        opexPct: { type: number }
        capexPct: { type: number }
        wbsCode: { type: string }
...
    Actual:
      type: object
      required:
        - id
        - projectId
        - projectName
        - year
        - month
        - actualResourceCost
        - actualVendorCost
        - actualInvoiceCost
        - source
      properties:
        id: { type: integer }
        projectId: { type: integer }
        projectName: { type: string }
        year: { type: integer }
        month: { type: integer }
        actualResourceCost: { type: number }
        actualVendorCost: { type: number }
        actualInvoiceCost: { type: number }
        source: { type: string, enum: [Manual, ExcelUpload, SAP, Timesheet] }
    ActualInput:
      type: object
      required:
        - projectId
        - year
        - month
        - actualResourceCost
        - actualVendorCost
        - actualInvoiceCost
        - source
      properties:
        projectId: { type: integer }
        year: { type: integer }
        month: { type: integer }
        actualResourceCost: { type: number }
        actualVendorCost: { type: number }
        actualInvoiceCost: { type: number }
        source: { type: string, enum: [Manual, ExcelUpload, SAP, Timesheet] }
...
    SignupInput:
      type: object
      required: [email, password, name]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
        name: { type: string }
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    MeResponse:
      type: object
      required: [user]
      properties:
        user: { $ref: "#/components/schemas/User" }
    User:
      type: object
      required: [id, email, name, role, createdAt]
      properties:
        id: { type: integer }
        email: { type: string }
        name: { type: string }
        role:
          type: string
          enum: [Admin, "PMO Lead", "Project Manager", "Finance Controller", "Resource Manager", "Demand Owner", Viewer]
        createdAt: { type: string, format: date-time }
        functionId: { type: integer, nullable: true }
        functionName: { type: string, nullable: true }
...